Add counting cost of road for class Car

+++ src/Practice_4_1_abstract/Vehicle.java
@@ -2,7 +2,7 @@ package Practice_4_1_abstract;
 
 public abstract class Vehicle {
     protected int distance;
-    protected float speed;
+    protected double speed;
     abstract void CostOfRoad();
     public Vehicle(int distance, float speed) {
         this.distance = distance;
@@ -17,7 +17,7 @@ public abstract class Vehicle {
     public void setSpeed(float speed) {
         this.speed = speed;
     }
-    public float getSpeed() {
+    public double getSpeed() {
         return speed;
     }
 }
